use auto generated package id

+++ src/database/entities/package_metadata.rs
@@ -6,8 +6,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
 #[sea_orm(table_name = "package_metadata")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: i64,
+    #[sea_orm(primary_key)]
+    pub id: i32,
     pub name: String,
     pub version: String,
     pub maintainer: String,
